Move search and replace to it's own file
load as if it was an extension.
